move cta-widget into /applications/static-pages
