+++ src/applications/static-pages/cta-widget/components/CallToActionAlert.jsx
