+++ src/applications/static-pages/cta-widget/components/messages/Verify.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import CallToActionAlert from './../CallToActionAlert';
+import CallToActionAlert from '../CallToActionAlert';
 
 const Verify = ({ serviceDescription, primaryButtonHandler }) => {
   const content = {
