+++ src/applications/static-pages/cta-widget/constants.js
