+++ src/applications/static-pages/cta-widget/ctaWidgets.js
