+++ src/applications/static-pages/cta-widget/index.js
