+++ src/applications/static-pages/cta-widget/tests/index.unit.spec.js
