+++ src/applications/static-pages/health-care-manage-benefits/refill-track-prescriptions-page/components/UnauthContent/index.js
@@ -4,7 +4,7 @@ import Telephone, {
   CONTACTS,
 } from '@department-of-veterans-affairs/component-library/Telephone';
 // Relative imports.
-import CallToActionWidget from 'platform/site-wide/cta-widget';
+import CallToActionWidget from 'applications/static-pages/cta-widget';
 
 export const UnauthContent = () => (
   <>
