+++ src/applications/static-pages/subscription-creators/subscribeAccordionEvents.js
