+++ src/applications/static-pages/subscription-creators/subscribeAdditionalInfoEvents.js
