+++ src/applications/static-pages/widget-creators/alerts-dismiss-view.js
