+++ src/applications/static-pages/widget-creators/createApplicationStatus.js
