+++ src/applications/static-pages/widget-creators/createMyVALoginWidget.js
