+++ src/applications/static-pages/widget-creators/ics-generator.js
