+++ src/applications/static-pages/widget-creators/resources-and-support-search.js
