+++ src/applications/static-pages/widget-creators/social-share-links.js
